Fixes issue running docker in docker (#26)
When running docker inside docker (dind), the value in /proc/1/cpuset
which is used to determine the container id is different. In this case,
the value is basically doubled, looking like
`/docker/{outer-id}/docker/{inner-id}`. This code only cares about the
most specific (right-most) id, whether running dind or not, for the
purposes of identify the worker container.