Feature/data dictionary register dataset (#57)
* POC of using endpoint to register a dataset in one request, using original Elwood annotations schema. Needs to revisit annotations file type, format, shape.

* Working on script to parse csv and send to api with correct elwood schema format

* Dojo API endpoint to post annotations from csv file. Integrating with UI next.

* Drive-by bugfix: we were requiring date format on an epoch date_type. The field wasnt being rendered but silently disallowed submissiong as the form had inline required validation errors

* Integrated UI to annotation dictionary file approach. Fixes a couple of bugs
along the way. Added some tests.

* Dataset register annotation: display loading animation when uploading and
parsing dictionary csv annotations

* Returning annotation dictionary errors from backend. Displaying these in
frontend. Annotations for fields not in dataset are not loaded by UI.
We can and should identify opportunities of how to validate the
fact that annotations for fields not in the dataset are not present, but API
allows for creation of dataset and upload annotations before a data file is uploaded.

* Added download csv template feature. Added error handling to uploading non csv
files for data dictionary.

* Added multi-column group csv-file annotation.

* Some cleanup and error handling for annotations for fields that dont exist in
UI. In the future we should figure out a way to also validate in the backend,
but not having the preview data makes this tricky.

* Allowing null/None values to be passed in from csv dictionary to es/elwood,
since elwood expects some keys event if they are empty. Handling an additional
scenario, still, where elwood tries to retrieve an undefined key in some
circumstances. It is optional and we shouldnt try to retrieve non existing keys.
Format validation errors for dictionary csv on UI. Added an Upload Annotation
File Dialog. Better loading of messaging when uploading dictionary file
annotation. Fixing a bug where the UI would be in an inconsistent state when
loading multi-part date annotations of columns that dont exist.

* Formatting of UI validation errors for dictionary files- was missing
a period and test was indicating a failure. Corrected.

* Moved various uploadFile definitions into client/utils in order to reuse.
slightly modified error message when uploading dictionary files. Loading and
errors are now reported within the Upload Annotation File dialog.

* Only allow uploading CSV annotations when not on edit metadata flow (ie when
addingAnnotationsAllowed is true on Flow.js config).

* Download csv/xls template from server to prepare for use of spreadsheet instead
of csv template, as well as other features we might add to template itself.

* Add more robust example Excel template file

* Able to download xlsx file for download template, for dataset file-dictionary registration.

* Cleanup, added docstrings, as well as added more helper text to dataset dictionary example file fields

* Allow both xslx spreadsheet OR csv dictionary files when annotating a dataset through UI.

---------

Co-authored-by: Matthew Printz <matt@jataware.com>